refactor: rearrange tests
move tests around
