+++ tests/clients/test_blinkstick_matrix.py
